Use config fragments instead of monolithic Kconfig
Until now, I've had a big qemu.config file that I've modified as needed.
The only way to find out what config options were important was to look
through the Git history for that file. This commit adds a few scripts
and files that changes over to a system where important config options
are specified explicitly and merged into one Kconfig file. This makes it
much easier to figure out what's being set and why.

merge_config.py is similar to scripts/kconfig/merge_config.sh in
linux.git, but it's slow and noisy. I might consolidate this in the
future.

I'm keeping the qemu.config file around for convenience. It should
always be the latest generated Kconfig.