btrfs_extent_tree_du: fix reflink double-accounting
If a single extent is referenced from the same file (or subvolume in -r
mode) multiple times, we will account it each time. Since we handle
every ref for a given extent before moving to the next extent, we can
fix it by caching the previous extent that we accounted for a given hash
table entry; if it matches the extent that we're currently accounting,
then don't account it again.