vm.py: don't use build directory for paths starting with / or .
`vm.py -k .` confusingly tries to run the builds directory as a kernel
rather than the current directory. This is because we treat any path
that doesn't contain a "/" as relative to the builds directory. If the
path starts with a ".", then the user clearly means relative to their
working directory, so don't look in the builds directory for that case.
Additionally, there's no reason the builds directory can't contain a
hierarchy, so only check for a "/" at the beginning (i.e., an absolute
path).