kconfig.py: add command to disable checking some options
metaconfig.py used to have a distinction between the defconfig and
fragments, and it ignored options from the defconfig that didn't make it
into the final config. This is useful because the defconfig contains
lots of config options that were copied from a distro config and we
don't really care about. Instead of reintroducing an idea of defconfig
to kconfig.py, add a way to explicitly disable checking for a region of
a config file.